move container viewed condition script and fix namespace and assembly
