+++ Assets/Scripts/SS3D/Networking/ObserverConditions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6a4ae4ac6b2ee64186554e60f408476
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
