+++ Assets/Scripts/SS3D/Networking/SS3D.Networking.asmdef
@@ -13,7 +13,8 @@
         "GUID:6aabbee860c32c64d9bbd52d0704ed2c",
         "GUID:75469ad4d38634e559750d17036d5f7c",
         "GUID:eb124556ba20d60469be9771eb1f0360",
-        "GUID:46e20dca70571ce4e817a6b5874168a0"
+        "GUID:46e20dca70571ce4e817a6b5874168a0",
+        "GUID:1476e4af5b8d89d4c9963efe314428b9"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
